requirements.txt
# discord==2.4.0
discord.py==2.4.0
psutil==6.0.0
aiohttp==3.10.5
